feat) opspec 0.1.6 support pkg icons fix) Markdown links to pkg content don't resolve properly; closes #360
