--- webapp/src/Pkg/Description.test.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import Description from './Description';
-
-it('renders without crashing', () => {
-  /* arrange */
-  const div = document.createElement('div');
-
-  /* act/assert */
-  ReactDOM.render(<Description value={'dummyDescription'}/>, div);
-});
