+++ webapp/src/Pkg/Input/Description.test.js
@@ -2,10 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import Description from './Description';
 
-it('renders without crashing', () => {
-  /* arrange */
-  const div = document.createElement('div');
+describe('value not null or empty', () => {
+  it('renders without crashing', () => {
+    /* arrange */
+    const div = document.createElement('div');
 
-  /* act/assert */
-  ReactDOM.render(<Description value={'dummyDescription'}/>, div);
+    /* act/assert */
+    ReactDOM.render(<Description value={'dummyDescription'}/>, div);
+  });
 });
