+++ webapp/src/Pkg/Output/Description.test.js
@@ -2,15 +2,12 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import Description from './Description';
 
-it('renders without crashing', () => {
-  /* arrange */
-  const div = document.createElement('div');
+describe('value not null or empty', () => {
+  it('renders without crashing', () => {
+    /* arrange */
+    const div = document.createElement('div');
 
-  /* act/assert */
-  ReactDOM.render(
-    <Description
-      pkgRef={'dummyPkgRef'}
-      value={'dummyValue'}
-    />,
-    div);
+    /* act/assert */
+    ReactDOM.render(<Description value={'dummyDescription'}/>, div);
+  });
 });
